Split the giant redash.controllers module into a package
